requirements_restore.txt
aistudio-sdk==0.3.8
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bce-python-sdk==0.9.46
blinker==1.9.0
cachetools==6.2.4
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
cssselect==1.3.0
cssutils==2.11.1
easyocr==1.7.2
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
filelock==3.19.1
Flask==3.1.2
fsspec==2025.10.0
ftfy==6.3.1
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.1
idna==3.10
imageio==2.37.0
imagesize==1.4.1
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
lazy_loader==0.4
lxml==6.0.2
MarkupSafe==2.1.5
modelscope==1.30.0
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.2.1
ninja==1.13.0
numpy<2.0.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
outcome==1.3.0.post0
paddleocr==3.2.0
paddlepaddle==3.2.0
paddlex==3.2.1
pandas==2.2.2
pefile==2023.2.7
pillow==10.3.0
premailer==3.10.0
prettytable==3.16.0
protobuf==3.20.3
psutil==7.1.0
py-cpuinfo==9.0.0
pyarmor==9.1.9
pyarmor.cli.core==7.6.8
pyarmor_runtime_000000==0.0.2
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyMuPDF==1.26.5
pypdfium2==4.30.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.1
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dotenv==1.2.1
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
regex==2025.11.3
requests==2.32.5
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
safetensors==0.6.2
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
selenium==4.36.0
shapely==2.0.7
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2024.8.30
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
trio==0.31.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
wcwidth==0.2.14
webdriver-manager==4.0.2
websocket-client==1.9.0
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.23.0